requirements.txt
transformers
torch 
torchvision
torchaudiojittor
numpy
collections
